Refactor save listing logic into own component
